keycloak-iam-system/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
requests==2.31.0
Werkzeug==2.3.7
cryptography==41.0.7
python-dotenv==1.0.0